+++ tools/src/com.oracle.truffle.tools.profiler/src/com/oracle/truffle/tools/profiler/impl/CPUTracerInstrument.java
@@ -86,24 +86,6 @@ public class CPUTracerInstrument extends TruffleInstrument {
      * Does a lookup in the runtime instruments of the engine and returns an instance of the
      * {@link CPUTracer}.
      *
-     * @since 0.30
-     * @deprecated use {@link #getTracer(Engine)} instead.
-     */
-    @Deprecated
-    @SuppressWarnings("deprecation")
-    public static CPUTracer getTracer(com.oracle.truffle.api.vm.PolyglotEngine engine) {
-        com.oracle.truffle.api.vm.PolyglotRuntime.Instrument instrument = engine.getRuntime().getInstruments().get(ID);
-        if (instrument == null) {
-            throw new IllegalStateException("Tracer is not installed.");
-        }
-        instrument.setEnabled(true);
-        return instrument.lookup(CPUTracer.class);
-    }
-
-    /**
-     * Does a lookup in the runtime instruments of the engine and returns an instance of the
-     * {@link CPUTracer}.
-     *
      * @since 0.33
      */
     public static CPUTracer getTracer(Engine engine) {